Extracted api.js to its own package kinto-client
