+++ .npmignore
@@ -1,2 +1,4 @@
 /.venv
 /src
+/esdoc
+/test
