+++ src/utils.js
@@ -21,26 +21,6 @@ export function deepEquals(a, b) {
 }
 
 /**
- * Returns the specified string with double quotes.
- *
- * @param  {String} str  A string to quote.
- * @return {String}
- */
-export function quote(str) {
-  return `"${str}"`;
-}
-
-/**
- * Trim double quotes from specified string.
- *
- * @param  {String} str  A string to unquote.
- * @return {String}
- */
-export function unquote(str) {
-  return str.replace(/^"/, "").replace(/"$/, "");
-}
-
-/**
  * Checks if a value is undefined.
  * @param  {Any}  value
  * @return {Boolean}
@@ -104,27 +84,6 @@ export function filterObjects(filters, list) {
 export function reduceRecords(filters, order, list) {
   const filtered = filters ? filterObjects(filters, list) : list;
   return order ? sortObjects(order, filtered) : filtered;
-}
-
-/**
- * Chunks an array into n pieces.
- *
- * @param  {Array}  array
- * @param  {Number} n
- * @return {Array}
- */
-export function partition(array, n) {
-  if (n <= 0) {
-    return array;
-  }
-  return array.reduce((acc, x, i) => {
-    if (i === 0 || i % n === 0) {
-      acc.push([x]);
-    } else {
-      acc[acc.length - 1].push(x);
-    }
-    return acc;
-  }, []);
 }
 
 /**
