+++ test/integration_test.js
@@ -7,7 +7,7 @@ import chai, { expect } from "chai";
 import chaiAsPromised from "chai-as-promised";
 import sinon from "sinon";
 import Kinto from "../src";
-import { cleanRecord } from "../src/api";
+import { cleanRecord } from "../src/collection";
 
 chai.use(chaiAsPromised);
 chai.should();
